Restructurize js files
