+++ saleor/static/js/mobile/filters-menu.js
@@ -1,0 +1,10 @@
+export default $(document).ready((e) => {
+  $('.filters-menu').on('click', (e) => {
+    const t = $('.filters-menu__body');
+    if (t.hasClass('d-none')) {
+      t.removeClass('d-none');
+    } else {
+      t.addClass('d-none');
+    }
+  });
+});
